Moved Menu and RefreshButton logic to their own files
