+++ packages/tools/devtools/devtools-view/src/DevtoolsView.tsx
@@ -3,15 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import {
-	Button,
-	FluentProvider,
-	Tooltip,
-	makeStyles,
-	shorthands,
-	tokens,
-} from "@fluentui/react-components";
-import { ArrowSync24Regular } from "@fluentui/react-icons";
+import { FluentProvider, makeStyles, shorthands, tokens } from "@fluentui/react-components";
 import { type ITelemetryBaseLogger } from "@fluidframework/core-interfaces";
 import {
 	type ContainerKey,
@@ -29,22 +21,16 @@ import { createChildLogger } from "@fluidframework/telemetry-utils/internal";
 import React from "react";
 
 import { useMessageRelay } from "./MessageRelayContext.js";
-import {
-	ConsoleVerboseLogger,
-	LoggerContext,
-	TelemetryOptInLogger,
-	useLogger,
-} from "./TelemetryUtils.js";
+import { ConsoleVerboseLogger, LoggerContext, TelemetryOptInLogger } from "./TelemetryUtils.js";
 import { ThemeContext, getFluentUIThemeToUse } from "./ThemeHelper.js";
 import {
 	ContainerDevtoolsView,
 	LandingView,
-	MenuItem,
-	MenuSection,
+	Menu,
 	NoDevtoolsErrorBar,
 	OpLatencyView,
-	TelemetryConsentModal,
 	SettingsView,
+	TelemetryConsentModal,
 	TelemetryView,
 	Waiting,
 } from "./components/index.js";
@@ -445,259 +431,4 @@ function View(props: ViewProps): React.ReactElement {
 	}
 
 	return <div className={styles.root}>{view}</div>;
-}
-
-const useMenuStyles = makeStyles({
-	root: {
-		...shorthands.gap("0px", "10px"),
-		...shorthands.padding("10px"),
-		"boxSizing": "border-box",
-		"display": "flex",
-		"flexDirection": "column",
-		"height": "100%",
-		"overflowY": "auto",
-		"minWidth": "150px",
-		// Ensures the last div/component is anchored to the bottom.
-		"> :last-child": {
-			marginTop: "auto",
-			marginBottom: "15px",
-		},
-	},
-
-	// TODO: dedupe with MenuItem
-	button: {
-		"alignItems": "center",
-		"cursor": "pointer",
-		"display": "flex",
-		"flexDirection": "row",
-		"paddingLeft": "5px",
-		"&:hover": {
-			color: tokens.colorNeutralForeground1Hover,
-			backgroundColor: tokens.colorNeutralBackground1Hover,
-		},
-	},
-});
-
-/**
- * {@link Menu} input props.
- */
-interface MenuProps {
-	/**
-	 * The current menu selection (if any).
-	 */
-	currentSelection?: MenuSelection;
-
-	/**
-	 * Sets the menu selection to the specified value.
-	 *
-	 * @remarks Passing `undefined` clears the selection.
-	 */
-	setSelection(newSelection: MenuSelection | undefined): void;
-
-	/**
-	 * Set of features supported by the {@link @fluidframework/devtools-core#IFluidDevtools}
-	 * instance being used by the page application.
-	 */
-	supportedFeatures: DevtoolsFeatureFlags;
-
-	/**
-	 * The set of Containers to offer as selection options.
-	 */
-	containers?: ContainerKey[];
-}
-
-/**
- * Menu component for {@link DevtoolsView}.
- */
-function Menu(props: MenuProps): React.ReactElement {
-	const { currentSelection, setSelection, supportedFeatures, containers } = props;
-	const usageLogger = useLogger();
-
-	const styles = useMenuStyles();
-
-	function onContainerClicked(containerKey: ContainerKey): void {
-		setSelection({ type: "containerMenuSelection", containerKey });
-		usageLogger?.sendTelemetryEvent({
-			eventName: "Navigation",
-			details: { target: "Menu_Container" },
-		});
-	}
-
-	function onTelemetryClicked(): void {
-		setSelection({ type: "telemetryMenuSelection" });
-		usageLogger?.sendTelemetryEvent({
-			eventName: "Navigation",
-			details: { target: "Menu_Telemetry" },
-		});
-	}
-
-	function onSettingsClicked(): void {
-		setSelection({ type: "settingsMenuSelection" });
-		usageLogger?.sendTelemetryEvent({
-			eventName: "Navigation",
-			details: { target: "Menu_Settings" },
-		});
-	}
-
-	function onHomeClicked(): void {
-		setSelection({ type: "homeMenuSelection" });
-		usageLogger?.sendTelemetryEvent({
-			eventName: "Navigation",
-			details: { target: "Menu_Home" },
-		});
-	}
-
-	function onOpLatencyClicked(): void {
-		setSelection({ type: "opLatencyMenuSelection" });
-		usageLogger?.sendTelemetryEvent({
-			eventName: "Navigation",
-			details: { target: "Menu_OpLatency" },
-		});
-	}
-
-	const menuSections: React.ReactElement[] = [];
-
-	menuSections.push(
-		<MenuSection header="Home" key="home-menu-section" onHeaderClick={onHomeClicked} />,
-		<ContainersMenuSection
-			key="containers-menu-section"
-			containers={containers}
-			currentContainerSelection={
-				currentSelection?.type === "containerMenuSelection"
-					? currentSelection.containerKey
-					: undefined
-			}
-			selectContainer={onContainerClicked}
-		/>,
-	);
-
-	// Display the Telemetry menu section only if the corresponding Devtools instance supports telemetry messaging.
-	if (supportedFeatures.telemetry === true) {
-		menuSections.push(
-			<MenuSection header="Telemetry" key="telemetry-menu-section">
-				<MenuItem
-					isActive={currentSelection?.type === "telemetryMenuSelection"}
-					text="Events"
-					onClick={onTelemetryClicked}
-				/>
-			</MenuSection>,
-		);
-	}
-
-	if (supportedFeatures.opLatencyTelemetry === true) {
-		menuSections.push(
-			<MenuSection
-				header="Op Latency"
-				key="op-latency-menu-section"
-				onHeaderClick={onOpLatencyClicked}
-			/>,
-		);
-	}
-
-	menuSections.push(
-		<MenuSection
-			header="Settings"
-			key="settings-menu-section"
-			onHeaderClick={onSettingsClicked}
-		/>,
-	);
-
-	return (
-		<div className={styles.root}>{menuSections.length === 0 ? <Waiting /> : menuSections}</div>
-	);
-}
-
-/**
- * {@link ContainersMenuSection} input props.
- */
-interface ContainersMenuSectionProps {
-	/**
-	 * The set of Containers to offer as selection options.
-	 */
-	containers?: ContainerKey[];
-
-	/**
-	 * The currently selected Container key, if one is currently selected.
-	 */
-	currentContainerSelection: ContainerKey | undefined;
-
-	/**
-	 * Updates the Container selection to the specified key.
-	 *
-	 * @remarks Passing `undefined` clears the selection.
-	 */
-	selectContainer(containerKey: ContainerKey | undefined): void;
-}
-
-/**
- * Displays the Containers menu section, allowing the user to select the Container to display.
- *
- * @remarks Displays a spinner while the Container list is being loaded (if the list is undefined),
- * and displays a note when there are no registered Containers (if the list is empty).
- */
-function ContainersMenuSection(props: ContainersMenuSectionProps): React.ReactElement {
-	const { containers, selectContainer, currentContainerSelection } = props;
-	let containerSectionInnerView: React.ReactElement;
-	if (containers === undefined) {
-		containerSectionInnerView = <Waiting label="Fetching Container list" />;
-	} else if (containers.length === 0) {
-		containerSectionInnerView = <div>No Containers found.</div>;
-	} else {
-		containers.sort((a: string, b: string) => a.localeCompare(b));
-		containerSectionInnerView = (
-			<>
-				{containers.map((containerKey: string) => (
-					<MenuItem
-						key={containerKey}
-						isActive={currentContainerSelection === containerKey}
-						text={containerKey}
-						onClick={(event): void => {
-							selectContainer(`${containerKey}`);
-						}}
-					/>
-				))}
-			</>
-		);
-	}
-
-	return (
-		<MenuSection
-			header="Containers"
-			key="container-selection-menu-section"
-			icon={<RefreshButton />}
-		>
-			{containerSectionInnerView}
-		</MenuSection>
-	);
-}
-
-/**
- * A refresh button to retrieve the latest list of containers.
- */
-function RefreshButton(): React.ReactElement {
-	const messageRelay = useMessageRelay();
-	const usageLogger = useLogger();
-
-	const transparentButtonStyle = {
-		backgroundColor: "transparent",
-		border: "none",
-		cursor: "pointer",
-	};
-
-	function handleRefreshClick(): void {
-		// Query for list of Containers
-		messageRelay.postMessage(getContainerListMessage);
-		usageLogger?.sendTelemetryEvent({ eventName: "ContainerRefreshButtonClicked" });
-	}
-
-	return (
-		<Tooltip content="Refresh Containers list" relationship="label">
-			<Button
-				icon={<ArrowSync24Regular />}
-				style={transparentButtonStyle}
-				onClick={handleRefreshClick}
-				aria-label="Refresh Containers list"
-			></Button>
-		</Tooltip>
-	);
 }
