+++ packages/tools/devtools/devtools-view/src/components/Menu.tsx
@@ -3,8 +3,15 @@
  * Licensed under the MIT License.
  */
 
-import { makeStyles, mergeClasses, tokens } from "@fluentui/react-components";
+import { makeStyles, mergeClasses, shorthands, tokens } from "@fluentui/react-components";
+import {
+	type HasContainerKey,
+	type DevtoolsFeatureFlags,
+	type ContainerKey,
+} from "@fluidframework/devtools-core/internal";
 import React from "react";
+import { useLogger } from "../TelemetryUtils.js";
+import { RefreshButton, Waiting } from "./index.js";
 
 /**
  * Props for {@link MenuSection}
@@ -26,6 +33,37 @@ export type MenuSectionProps = React.PropsWithChildren<{
 	onHeaderClick?(): void;
 }>;
 
+const useMenuStyles = makeStyles({
+	root: {
+		...shorthands.gap("0px", "10px"),
+		...shorthands.padding("10px"),
+		"boxSizing": "border-box",
+		"display": "flex",
+		"flexDirection": "column",
+		"height": "100%",
+		"overflowY": "auto",
+		"minWidth": "150px",
+		// Ensures the last div/component is anchored to the bottom.
+		"> :last-child": {
+			marginTop: "auto",
+			marginBottom: "15px",
+		},
+	},
+
+	// TODO: dedupe with MenuItem
+	button: {
+		"alignItems": "center",
+		"cursor": "pointer",
+		"display": "flex",
+		"flexDirection": "row",
+		"paddingLeft": "5px",
+		"&:hover": {
+			color: tokens.colorNeutralForeground1Hover,
+			backgroundColor: tokens.colorNeutralBackground1Hover,
+		},
+	},
+});
+
 const useMenuSectionStyles = makeStyles({
 	root: {
 		display: "flex",
@@ -39,6 +77,74 @@ const useMenuSectionStyles = makeStyles({
 		cursor: "pointer",
 	},
 });
+
+/**
+ * Indicates that the currently selected menu option is a particular Container.
+ * @see {@link MenuSection} for other possible options.
+ */
+interface ContainerMenuSelection extends HasContainerKey {
+	/**
+	 * String to differentiate between different types of options in menu.
+	 */
+	type: "containerMenuSelection";
+}
+
+/**
+ * Indicates that the currently selected menu option is the Telemetry view.
+ * @see {@link MenuSection} for other possible options.
+ */
+interface TelemetryMenuSelection {
+	/**
+	 * String to differentiate between different types of options in menu.
+	 */
+	type: "telemetryMenuSelection";
+}
+
+/**
+ * Indicates that the currently selected menu option is the Settings view.
+ * @see {@link MenuSection} for other possible options.
+ */
+interface SettingsMenuSelection {
+	/**
+	 * String to differentiate between different types of options in menu.
+	 */
+	type: "settingsMenuSelection";
+}
+
+/**
+ * Indicates that the currently selected menu option is the Home view.
+ * @see {@link MenuSection} for other possible options.
+ */
+interface HomeMenuSelection {
+	/**
+	 * String to differentiate between different types of options in menu.
+	 */
+	type: "homeMenuSelection";
+}
+
+/**
+ * Indicates that the currently selected menu option is the Op Latency view
+ * @see {@link MenuSection} for other possible options.
+ */
+interface OpLatencyMenuSelection {
+	/**
+	 * String to differentiate between different types of options in menu.
+	 */
+	type: "opLatencyMenuSelection";
+}
+
+/**
+ * Discriminated union type for all the selectable options in the menu.
+ * Each specific type should contain any additional information it requires.
+ * E.g. {@link ContainerMenuSelection} represents that the menu option for a Container
+ * is selected, and has a 'containerKey' property to indicate which Container.
+ */
+type MenuSelection =
+	| TelemetryMenuSelection
+	| ContainerMenuSelection
+	| SettingsMenuSelection
+	| HomeMenuSelection
+	| OpLatencyMenuSelection;
 
 /**
  * Generic component for a section of the menu.
@@ -103,5 +209,197 @@ export function MenuItem(props: MenuItemProps): React.ReactElement {
 		<div className={style} onClick={onClick}>
 			{text}
 		</div>
+	);
+}
+
+/**
+ * {@link Menu} input props.
+ */
+interface MenuProps {
+	/**
+	 * The current menu selection (if any).
+	 */
+	currentSelection?: MenuSelection;
+
+	/**
+	 * Sets the menu selection to the specified value.
+	 *
+	 * @remarks Passing `undefined` clears the selection.
+	 */
+	setSelection(newSelection: MenuSelection | undefined): void;
+
+	/**
+	 * Set of features supported by the {@link @fluidframework/devtools-core#IFluidDevtools}
+	 * instance being used by the page application.
+	 */
+	supportedFeatures: DevtoolsFeatureFlags;
+
+	/**
+	 * The set of Containers to offer as selection options.
+	 */
+	containers?: ContainerKey[];
+}
+/**
+ * {@link ContainersMenuSection} input props.
+ */
+interface ContainersMenuSectionProps {
+	/**
+	 * The set of Containers to offer as selection options.
+	 */
+	containers?: ContainerKey[];
+
+	/**
+	 * The currently selected Container key, if one is currently selected.
+	 */
+	currentContainerSelection: ContainerKey | undefined;
+
+	/**
+	 * Updates the Container selection to the specified key.
+	 *
+	 * @remarks Passing `undefined` clears the selection.
+	 */
+	selectContainer(containerKey: ContainerKey | undefined): void;
+}
+
+/**
+ * Displays the Containers menu section, allowing the user to select the Container to display.
+ *
+ * @remarks Displays a spinner while the Container list is being loaded (if the list is undefined),
+ * and displays a note when there are no registered Containers (if the list is empty).
+ */
+function ContainersMenuSection(props: ContainersMenuSectionProps): React.ReactElement {
+	const { containers, selectContainer, currentContainerSelection } = props;
+	let containerSectionInnerView: React.ReactElement;
+	if (containers === undefined) {
+		containerSectionInnerView = <Waiting label="Fetching Container list" />;
+	} else if (containers.length === 0) {
+		containerSectionInnerView = <div>No Containers found.</div>;
+	} else {
+		containers.sort((a: string, b: string) => a.localeCompare(b));
+		containerSectionInnerView = (
+			<>
+				{containers.map((containerKey: string) => (
+					<MenuItem
+						key={containerKey}
+						isActive={currentContainerSelection === containerKey}
+						text={containerKey}
+						onClick={(event): void => {
+							selectContainer(`${containerKey}`);
+						}}
+					/>
+				))}
+			</>
+		);
+	}
+
+	return (
+		<MenuSection
+			header="Containers"
+			key="container-selection-menu-section"
+			icon={<RefreshButton />}
+		>
+			{containerSectionInnerView}
+		</MenuSection>
+	);
+}
+
+/**
+ * Menu component for {@link DevtoolsView}.
+ */
+export function Menu(props: MenuProps): React.ReactElement {
+	const { currentSelection, setSelection, supportedFeatures, containers } = props;
+	const usageLogger = useLogger();
+
+	const styles = useMenuStyles();
+
+	function onContainerClicked(containerKey: ContainerKey): void {
+		setSelection({ type: "containerMenuSelection", containerKey });
+		usageLogger?.sendTelemetryEvent({
+			eventName: "Navigation",
+			details: { target: "Menu_Container" },
+		});
+	}
+
+	function onTelemetryClicked(): void {
+		setSelection({ type: "telemetryMenuSelection" });
+		usageLogger?.sendTelemetryEvent({
+			eventName: "Navigation",
+			details: { target: "Menu_Telemetry" },
+		});
+	}
+
+	function onSettingsClicked(): void {
+		setSelection({ type: "settingsMenuSelection" });
+		usageLogger?.sendTelemetryEvent({
+			eventName: "Navigation",
+			details: { target: "Menu_Settings" },
+		});
+	}
+
+	function onHomeClicked(): void {
+		setSelection({ type: "homeMenuSelection" });
+		usageLogger?.sendTelemetryEvent({
+			eventName: "Navigation",
+			details: { target: "Menu_Home" },
+		});
+	}
+
+	function onOpLatencyClicked(): void {
+		setSelection({ type: "opLatencyMenuSelection" });
+		usageLogger?.sendTelemetryEvent({
+			eventName: "Navigation",
+			details: { target: "Menu_OpLatency" },
+		});
+	}
+
+	const menuSections: React.ReactElement[] = [];
+
+	menuSections.push(
+		<MenuSection header="Home" key="home-menu-section" onHeaderClick={onHomeClicked} />,
+		<ContainersMenuSection
+			key="containers-menu-section"
+			containers={containers}
+			currentContainerSelection={
+				currentSelection?.type === "containerMenuSelection"
+					? currentSelection.containerKey
+					: undefined
+			}
+			selectContainer={onContainerClicked}
+		/>,
+	);
+
+	// Display the Telemetry menu section only if the corresponding Devtools instance supports telemetry messaging.
+	if (supportedFeatures.telemetry === true) {
+		menuSections.push(
+			<MenuSection header="Telemetry" key="telemetry-menu-section">
+				<MenuItem
+					isActive={currentSelection?.type === "telemetryMenuSelection"}
+					text="Events"
+					onClick={onTelemetryClicked}
+				/>
+			</MenuSection>,
+		);
+	}
+
+	if (supportedFeatures.opLatencyTelemetry === true) {
+		menuSections.push(
+			<MenuSection
+				header="Op Latency"
+				key="op-latency-menu-section"
+				onHeaderClick={onOpLatencyClicked}
+			/>,
+		);
+	}
+
+	menuSections.push(
+		<MenuSection
+			header="Settings"
+			key="settings-menu-section"
+			onHeaderClick={onSettingsClicked}
+		/>,
+	);
+
+	return (
+		<div className={styles.root}>{menuSections.length === 0 ? <Waiting /> : menuSections}</div>
 	);
 }
