+++ packages/tools/devtools/devtools-view/src/components/index.ts
@@ -17,10 +17,11 @@ export * from "./ContainerSummaryView.js";
 export * from "./LandingView.js";
 export * from "./Menu.js";
 export * from "./NoDevtoolsErrorBar.js";
+export * from "./OpLatencyView.js";
+export * from "./SettingsView.js";
+export * from "./RefreshButton.js";
 export * from "./TelemetryConsentModal.js";
 export * from "./TelemetryView.js";
-export * from "./SettingsView.js";
-export * from "./OpLatencyView.js";
 export * from "./Waiting.js";
 
 /* eslint-enable no-restricted-syntax */
